Move UInt32Tests
